Mise en place de l'outil de test end to end dans le projet client
