+++ client/.env
@@ -1,1 +1,2 @@
 HTTP_TEST_PORT=3000
+NODE_ENV=test
