--- client/cypress.json
@@ -1,37 +1,0 @@
-{
-  "projectId": "tdfyudugbuy11",
-  "baseUrl": "http://localhost:3000/",
-  "port": 9090,
-  "video": true,
-  "testFiles": "**/*.specs.ts",
-  "env": {
-    "cypress-movie": {
-      "enabled": true,
-      "cursorTracking": {
-        "enabled": false,
-        "shape": "arrow"
-      },
-      "width": 1920,
-      "height": 640
-    }
-  },
-  "reporter": "cypress-multi-reporters",
-  "reporterOptions": {
-    "reporterEnabled": "cypress-sonarqube-reporter, mochawesome",
-    "mochawesomeReporterOptions": {
-      "reportDir": "cypress/reports/tests",
-      "charts": true,
-      "html": true,
-      "json": true,
-      "reportPageTitle": "Browser Integration Tests Reports",
-      "embeddedScreenshots": true,
-      "inlineAssets": true,
-      "overwrite": true
-    },
-    "cypressSonarqubeReporterReporterOptions": {
-      "outputDir": "cypress/reports/sonar",
-      "useAbsoluteSpecPath": true
-    }
-  },
-  "screenshotsFolder": "cypress/reports/tests/assets"
-}
