+++ client/cypress/e2e/first.cy.ts
@@ -1,54 +1,65 @@
-/* eslint-disable promise/no-nesting */
-/* eslint-disable promise/catch-or-return */
-
+// type definitions for Cypress object "cy"
 /// <reference types="cypress" />
+
+// type definitions for custom commands like "createDefaultTodos"
+/// <reference types="../src/support" />
+
+// check this file using TypeScript if available
+// @ts-check
 
 import { expect } from "chai";
 import specTitle from "cypress-sonarqube-reporter/specTitle";
-import { CyHttpMessages } from "cypress/types/net-stubbing";
-import { basicInputType } from "./components/BasicInputs";
-import { addEnd, addPresentation } from "./presentation";
+// import { CyHttpMessages } from "cypress/types/net-stubbing";
+
+describe(specTitle("Page Access"), () => {
+  it("Visite page", () => {
+    // cy.clearViewport();
+    cy.visit("/");
+    cy.addPresentation("Settings");
+    cy.get("body");
+  });
+});
 
 describe(specTitle("Offlines Access"), () => {
   it("Settings", () => {
-    cy.clearViewport();
+    // cy.clearViewport();
     cy.visit("/");
 
-    addPresentation("Settings");
+    cy.addPresentation("Settings");
 
     cy.wait(2000);
     cy.visit("/");
     // all parameters are optional
-    cy.toast("Change language to French", {
+    /*cy.toast("Change language to French", {
       blocking: true
-    });
+    });*/
     cy.get("#lang-fr").click();
     cy.contains("Connexion");
 
     // all parameters are optional
-    cy.toast("Change language to English", {
+    /*cy.toast("Change language to English", {
       blocking: true
-    });
+    });*/
     cy.get("#lang-en").click();
     cy.contains("Sign In");
 
     cy.get("#PAGES_SETTINGS_TITLE").click();
 
     // all parameters are optional
-    cy.toast("Display menu", {
+    /*cy.toast("Display menu", {
       blocking: true
-    });
+    });*/
 
-    addEnd();
+    cy.addEnd();
   });
 });
 
 describe(specTitle("Subscription"), () => {
   it("With Success", () => {
-    cy.clearViewport();
+    // cy.clearViewport();
     cy.visit("/");
 
-    addPresentation("Subscription - with success");
+    cy.addPresentation("Subscription - with success");
 
     cy.wait(2000);
     cy.visit("/");
@@ -56,7 +67,10 @@ describe(specTitle("Subscription"), () => {
     cy.intercept(
       "POST",
       "/api/v1/user/sign-up",
-      (req: CyHttpMessages.IncomingHttpRequest) => {
+      (
+        req
+        // : CyHttpMessages.IncomingHttpRequest
+      ) => {
         expect(req.body?.email).to.equal("mguillaumef@draft.com");
         expect(req.body?.login).to.equal("mguillaumef");
         expect(req.body?.password).to.equal("DROWssap987");
@@ -66,27 +80,27 @@ describe(specTitle("Subscription"), () => {
 
     cy.get("#PAGES_SIGN_UP_LINK").click();
 
-    basicInputType("FORMS_SIGN_UP_FIELDS_EMAIL", "mguillaumef@draft.com");
+    cy.basicInputType("FORMS_SIGN_UP_FIELDS_EMAIL", "mguillaumef@draft.com");
 
-    basicInputType("FORMS_SIGN_UP_FIELDS_IDENTIFIER", "mguillaumef");
+    cy.basicInputType("FORMS_SIGN_UP_FIELDS_IDENTIFIER", "mguillaumef");
 
-    basicInputType("FORMS_SIGN_UP_FIELDS_PASSWORD", "DROWssap987");
+    cy.basicInputType("FORMS_SIGN_UP_FIELDS_PASSWORD", "DROWssap987");
 
-    basicInputType("FORMS_SIGN_UP_FIELDS_CONFIRM", "DROWssap987").type(
+    cy.basicInputType("FORMS_SIGN_UP_FIELDS_CONFIRM", "DROWssap987").type(
       "{enter}"
     );
 
     cy.wait("@subscribe");
-    addEnd();
+    cy.addEnd();
   });
 });
 
 describe(specTitle("Login"), () => {
   it("With Success", () => {
-    cy.clearViewport();
+    // cy.clearViewport();
     cy.visit("/");
 
-    addPresentation("Login - with success");
+    cy.addPresentation("Login - with success");
 
     cy.wait(2000);
     cy.visit("/");
@@ -94,7 +108,10 @@ describe(specTitle("Login"), () => {
     cy.intercept(
       "POST",
       "/api/v1/user/sign-in",
-      (req: CyHttpMessages.IncomingHttpRequest) => {
+      (
+        req
+        // : CyHttpMessages.IncomingHttpRequest
+      ) => {
         expect(req.body?.login).to.equal("mguillaumef");
         expect(req.body?.password).to.equal("DROWssap987");
       }
@@ -102,28 +119,28 @@ describe(specTitle("Login"), () => {
 
     cy.get("#PAGES_SIGN_IN_LINK").click();
 
-    basicInputType("FORMS_SIGN_IN_FIELDS_IDENTIFIER", "mguillaumef");
+    cy.basicInputType("FORMS_SIGN_IN_FIELDS_IDENTIFIER", "mguillaumef");
 
-    basicInputType("FORMS_SIGN_IN_FIELDS_PASSWORD", "DROWssap987").type(
+    cy.basicInputType("FORMS_SIGN_IN_FIELDS_PASSWORD", "DROWssap987").type(
       "{enter}"
     );
 
     cy.wait("@login");
-    addEnd();
+    cy.addEnd();
   });
 });
 
 describe(specTitle("Errors pages"), () => {
   it("Not Found", () => {
-    cy.clearViewport();
+    // cy.clearViewport();
     cy.visit("/");
 
-    addPresentation("Not Found");
+    cy.addPresentation("Not Found");
 
     cy.wait(2000);
     cy.visit("/badPage");
     cy.contains("Error404");
 
-    addEnd();
+    cy.addEnd();
   });
 });
