--- client/cypress/support/index.js
@@ -1,40 +1,0 @@
-// ***********************************************************
-// This example support/index.js is processed and
-// loaded automatically before your test files.
-//
-// This is a great place to put global configuration and
-// behavior that modifies Cypress.
-//
-// You can change the location of this file or turn off
-// automatically serving support files with the
-// 'supportFile' configuration option.
-//
-// You can read more here:
-// https://on.cypress.io/configuration
-// ***********************************************************
-
-// Import commands.js using ES2015 syntax:
-// Alternatively you can use CommonJS syntax:
-// require('./commands')
-// coverage support
-import "@cypress/code-coverage/support";
-import "cypress-movie/commands";
-import addContext from "mochawesome/addContext";
-import "./commands";
-
-Cypress.on("test:after:run", (test, runnable) => {
-  if (test.state === "failed") {
-    // `assets/${Cypress.spec.name}/${runnable.parent.title} -- ${test.title} (failed).png`;
-    const screenshot = `assets/${
-      Cypress.spec.name
-    }/${`${runnable.parent.title} -- ${test.title} (failed).png`.replace(
-      /^(.*)\s\[@spec:?\s\{.*\}\]\s(--\s.*\s\(failed\)\.png)$/,
-      "$1 $2"
-    )}`;
-    // `${Cypress.config("screenshotsFolder")}/${
-    //      Cypress.spec.name
-    //    }/${runnable.parent.title} -- ${test.title} (failed).png`;
-
-    addContext({ test }, screenshot);
-  }
-});
