+++ client/package.json
@@ -8,15 +8,19 @@
     "build": "webpack --define-process-env-node-env production --progress --color --env PUBLIC_PATH=\"'.'\"",
     "build:doc": "styleguidist build && typedoc --out ./dist/docs/ts ./src/index.tsx",
     "start": "webpack serve --open --progress --color --define-process-env-node-env development",
+    "cy:build": "cd cypress && tsc --project tsconfig.json",
+    "postcy:build": "mv cypress/cypress.config.js . || move cypress\\cypress.config.js .",
+    "precy:open": "npm run cy:build",
+    "precy:run": "npm run cy:build",
     "cy:open": "cypress open",
-    "cy:run": "wait-on http://localhost:3000/ && cypress run --config-file cypress.json",
+    "cy:run": "wait-on http://localhost:3000/ && cypress run",
     "prettier": "prettier --write . --ignore-path ./.gitignore",
     "lint:eslint": "eslint . --fix --format html --output-file ./dist/reports/lint/eslint-report.html --ignore-path ./.gitignore",
     "lint:eslint:json": "eslint . --fix --format json --output-file ./dist/reports/lint/eslint-report.json --ignore-path ./.gitignore",
     "lint:stylelint": "stylelint \"**/*.{css,scss,sass}\" --custom-syntax postcss-scss --fix --formatter json > ./dist/reports/lint/stylelint-report.json --ignore-path ./.gitignore",
     "unit:test": "nyc mocha --config test/runners/mocha/.mocharc.jsonc --reporter mocha-multi-reporters --reporter-options configFile=mocha-multi-config.json",
-    "smog:test": "npm-run-all -p start cy:run",
-    "test": "smog:test",
+    "e2e:test": "npm-run-all -p start cy:run",
+    "test": "npm run unit:test && npm run e2e:test",
     "webpack-stats": "webpack --define-process-env-node-env production --profile --json=stats.json"
   },
   "browserslist": {
@@ -49,6 +53,7 @@
     "@types/core-js": "^2.5.5",
     "@types/jest": "^29.2.5",
     "@types/mocha": "^10.0.1",
+    "@types/mochawesome": "^6.2.1",
     "@types/node": "^18.11.18",
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.10",
@@ -69,10 +74,11 @@
     "compression-webpack-plugin": "^10.0.0",
     "css-loader": "^6.7.3",
     "css-minimizer-webpack-plugin": "^4.2.2",
-    "cypress": "^9.7.0",
+    "cypress": "^12.2.0",
     "cypress-movie": "^3.0.20",
     "cypress-multi-reporters": "^1.6.2",
     "cypress-sonarqube-reporter": "^1.11.0",
+    "dotenv": "^16.0.3",
     "eslint": "^8.31.0",
     "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-cypress": "^2.12.1",
