+++ sonar-project.properties
@@ -7,8 +7,7 @@ sonar.projectName=hangman-game
 sonar.projectVersion=1.0.0
 
 # Path is relative to the sonar-project.properties file. Replace "\" by "/" on Windows.
-sonar.modules=client,server,data-management,tools/smog-test-runner
-
+sonar.modules=client,server,data-management
 # Encoding of the source code. Default is default system encoding
 sonar.sourceEncoding=UTF-8
 
@@ -21,5 +20,3 @@ sonar.cfamily.threads=1
 
 #sonar.host.url=YourSonarCloud/SonarQubeURL
 sonar.testExecutionReportPaths=server/reports/boost-test-sonarqube-report-UT.xml,server/reports/boost-test-sonarqube-report-IT.xml,data-management/reports/boost-test-sonarqube-report-UT.xml,client/cypress/reports/sonar/sonar-report-all.xml,client/dist/reports/test-results.xml
-# TODO reset when dependency fixed
-# ,tools/smog-test-runner/cypress/reports/sonar/sonar-report-all.xml
