--- tools/smog-test-runner/cypress/src/support/e2e.ts
@@ -1,4 +1,0 @@
-// load type definitions that come with Cypress module
-/// <reference types="cypress" />
-import "@cypress/code-coverage/support";
-import "./commands";
