--- tools/smog-test-runner/package.json
@@ -1,50 +1,0 @@
-{
-  "author": "MGuillaumeF",
-  "name": "smog-test-runner",
-  "description": "tool to run smog tests of application",
-  "version": "0.1.0",
-  "private": true,
-  "main": "index.ts",
-  "scripts": {
-    "build": "cd cypress && tsc --project tsconfig.json",
-    "postbuild": "mv cypress/cypress.config.js . || move cypress\\cypress.config.js .",
-    "start": "cd ../../client;npm start",
-    "cy:open": "cypress open",
-    "cy:run": "wait-on http://localhost:3000/ && cypress run",
-    "prettier": "prettier --write . --ignore-path ./.gitignore",
-    "lint:eslint": "eslint . --fix --format html --output-file ./eslint-report.html --ignore-path ./.gitignore",
-    "lint:eslint:json": "eslint . --fix --format json --output-file ./eslint-report.json --ignore-path ./.gitignore",
-    "pretest": "npm run build",
-    "test": "npm-run-all -p start cy:run"
-  },
-  "devDependencies": {
-    "@cypress/code-coverage": "^3.10.0",
-    "@mguillaumef/audiso": "^0.0.4",
-    "@types/node": "^18.11.17",
-    "@typescript-eslint/eslint-plugin": "^5.47.0",
-    "@typescript-eslint/parser": "^5.47.0",
-    "chai": "^4.3.7",
-    "cypress": "^12.2.0",
-    "cypress-multi-reporters": "^1.6.2",
-    "dotenv": "^16.0.3",
-    "eslint": "^8.30.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-cypress": "^2.12.1",
-    "eslint-plugin-jest": "^27.1.7",
-    "eslint-plugin-jest-react": "^0.1.0",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-promise": "^6.1.1",
-    "eslint-plugin-react": "^7.31.11",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-perf": "^3.3.1",
-    "eslint-plugin-typescript-sort-keys": "^2.1.0",
-    "eslint-webpack-plugin": "^3.2.0",
-    "mocha": "^10.2.0",
-    "mochawesome": "^7.1.3",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^2.8.1",
-    "typescript": "^4.9.4",
-    "wait-on": "^7.0.1"
-  }
-}
