--- tools/smog-test-runner/sonar-project.properties
@@ -1,4 +1,0 @@
-sonar.sources=package.json
-sonar.tests=cypress/src/e2e
-
-sonar.externalIssuesReportPaths=audit-report.json
